Show how to use Math.random in different ways

+++ java/learning_3/Main.java
@@ -13,21 +13,34 @@ public class Main {
     // int result = (int)Math.floor(1.1F);
     //   System.out.println(result);
 
-    // int result = (int)Math.max(1, 2);
+    // int result = Math.max(1, 2);
     // System.out.println(result);
 
-    // int result = (int)Math.min(1, 2);
+    // int result = Math.min(1, 2);
     // System.out.println(result);
 
-    // double result = (int)Math.random() ;
+    // double result = Math.random() ;
     // System.out.println(result);
 
-    double result = (int)Math.random() * 100;
+    // double result = Math.random() * 100;
+    // System.out.println(result);
+
+
+
+    // if you want a decimal value
+    // double result = Math.round(Math.random() * 100);
+    // System.out.println(result);
+
+       
+
+    // if you want an integer
+
+    // int result = (int) Math.round(Math.random() * 100);
+    // System.out.println(result);
+       
+
+    int result = (int) (Math.random() * 100);
     System.out.println(result);
-
-       
-
-
